app/requirements.txt
langchain==0.1.8
langchain-community==0.0.19
langchain-ollama==0.0.1
chromadb==0.4.22
pyyaml==6.0.1
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1